Version 2.11 - Updates from Requests from Initial Pull Request
Added logic so that date spans greater than 30 days will now create
multiple requests (for example if 1/1/2016 to 2/1/2016 is added, it
would create two requests instead of 1). In addition, I would like to
keep the requirement to install packages. I'm using quite a few and if
they are not installed the extension will fail. I believe in this case
if the package is installed already it will default to it and NOT try to
update.