Version 2.11 - Updates based on Request of Last Pull Request
Added logic so that date spans greater than 30 days will now create
multiple requests (ex. if 1/1/2016 to 2/1/2016 is passed it would create
two requests since it spans over 30+ days instead of 1). In addition, I
would like to keep the requirement to install packagtes. I'm using quite
a few and if they are not installed the extension will fail. I believe
in this case if the packagte is isntalled already it willd efault to it
and NOT try to update the package.